SW-8817 Extract stratum plot-selection test helper functions
In preparation for changes to the plot selection algorithm that will
require their own test suite, move the test fixture setup functions out
of StratumModelTest into a new BaseStratumModelTest superclass.

No functional changes here, just code movement.
